# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "juny-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/wellness": {
      "post": {
        "tags": [
          "wellness"
        ],
        "summary": "Create Wellness Log",
        "description": "Create a wellness log entry for a host.",
        "operationId": "create_wellness_log_api_v1_wellness_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WellnessLogCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellnessLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "wellness"
        ],
        "summary": "List Wellness Logs",
        "description": "List wellness logs for a host (paginated, newest first).",
        "operationId": "list_wellness_logs_api_v1_wellness_get",
        "parameters": [
          {
            "name": "host_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Host Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WellnessLogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wellness/{log_id}": {
      "get": {
        "tags": [
          "wellness"
        ],
        "summary": "Get Wellness Log",
        "description": "Get a specific wellness log by ID.",
        "operationId": "get_wellness_log_api_v1_wellness__log_id__get",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellnessLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_WellnessLogResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WellnessLogResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PaginatedResponse[WellnessLogResponse]"
      },
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Items per page"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "Total number of pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "description": "Whether there is a next page"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev",
            "description": "Whether there is a previous page"
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "total_pages",
          "has_next",
          "has_prev"
        ],
        "title": "PaginationMeta",
        "description": "Pagination metadata."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WellnessLogCreate": {
        "properties": {
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "status": {
            "$ref": "#/components/schemas/WellnessStatus"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "host_id",
          "status",
          "summary"
        ],
        "title": "WellnessLogCreate",
        "description": "Request body for creating a wellness log."
      },
      "WellnessLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "host_id": {
            "type": "string",
            "format": "uuid",
            "title": "Host Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "host_id",
          "status",
          "summary",
          "details",
          "created_at"
        ],
        "title": "WellnessLogResponse",
        "description": "Response model for a wellness log."
      },
      "WellnessStatus": {
        "type": "string",
        "enum": [
          "normal",
          "warning",
          "emergency"
        ],
        "title": "WellnessStatus",
        "description": "Wellness log severity levels."
      }
    }
  }
}